# internal/swagger/swagger.yaml
definitions:
...
  github_com_tokenlive_tokenlive-admin_pkg_errors.Error:
    properties:
      code:
        type: integer
      detail:
        type: string
      id:
        type: string
      status:
        type: string
    type: object
...
  schema.DegradeConfig:
    properties:
      attributes:
        additionalProperties:
          type: string
        type: object
      response_body:
        type: string
      response_code:
        type: integer
    type: object
...
  schema.ErrorParserPolicy:
    properties:
      content_types:
        items:
          type: string
        type: array
      expression:
        type: string
      parser:
        type: string
      statuses:
        items:
          type: string
        type: array
    type: object
...
  schema.FallbackPolicy:
    properties:
      targets:
        description: 降级目标模型链条，如 ["gpt-4:free", "gpt-3.5-turbo"]
        items:
          type: string
        type: array
    type: object
...
  schema.PolicyCircuitBreak:
    properties:
      allowed_calls_in_half_open_state:
        type: integer
      code_policy:
        type: string
      created_at:
        type: string
      creator:
        type: string
      degrade_config:
        type: string
      description:
        type: string
      enabled:
        type: integer
      error_codes:
        type: string
      error_messages:
        type: string
      failure_rate_threshold:
        type: number
      force_open:
        type: integer
      id:
        type: string
      level:
        type: string
      message_policy:
        type: string
      min_calls_threshold:
        type: integer
      model_id:
        type: string
      modifier:
        type: string
      name:
        type: string
      outlier_max_percent:
        type: integer
      priority:
        type: integer
      scope_code:
        type: string
      scope_type:
        type: string
      sliding_window_size:
        type: integer
      sliding_window_type:
        type: string
      slow_call_duration_threshold:
        type: integer
      slow_call_metric:
        type: string
      slow_call_rate_threshold:
        type: number
      updated_at:
        type: string
      version:
        type: integer
      wait_duration_in_open_state:
        type: integer
    type: object
  schema.PolicyCircuitBreakForm:
    properties:
      allowed_calls_in_half_open_state:
        description: Allowed calls in half open state
        type: integer
      code_policy:
        allOf:
        - $ref: '#/definitions/schema.ErrorParserPolicy'
        description: Code policy
      created_at:
        description: Create timestamp
        type: string
      creator:
        description: Creator
        type: string
      degrade_config:
        allOf:
        - $ref: '#/definitions/schema.DegradeConfig'
        description: Degrade config
      description:
        description: Details
        type: string
      enabled:
        description: Enabled
        type: integer
      error_codes:
        description: Error codes
        items:
          type: string
        type: array
      error_messages:
        description: Error messages
        items:
          type: string
        type: array
      failure_rate_threshold:
        description: Failure rate threshold
        type: number
      force_open:
        description: Force open
        type: integer
      id:
        type: string
      level:
        description: Policy level
        maxLength: 64
        type: string
      message_policy:
        allOf:
        - $ref: '#/definitions/schema.ErrorParserPolicy'
        description: Message policy
      min_calls_threshold:
        description: Min calls threshold
        type: integer
      model_id:
        description: Owner model ID; empty means template
        maxLength: 20
        type: string
      modifier:
        description: Modifier
        type: string
      name:
        description: Policy name
        maxLength: 128
        type: string
      outlier_max_percent:
        description: Outlier max percent
        type: integer
      priority:
        description: Priority
        minimum: 0
        type: integer
      scope_code:
        description: Scope code
        maxLength: 128
        type: string
      scope_type:
        description: Scope type
        maxLength: 32
        type: string
      sliding_window_size:
        description: Sliding window size
        type: integer
      sliding_window_type:
        description: Sliding window type
        maxLength: 16
        type: string
      slow_call_duration_threshold:
        description: Slow call duration threshold
        type: integer
      slow_call_metric:
        description: Slow call metric
        type: string
      slow_call_rate_threshold:
        description: Slow call rate threshold
        type: number
      updated_at:
        description: Update timestamp
        type: string
      version:
        description: Version
        type: integer
      wait_duration_in_open_state:
        description: Wait duration in open state
        type: integer
    required:
    - level
    - name
    - sliding_window_type
    type: object
  schema.PolicyCopyToModelForm:
    properties:
      model_id:
        maxLength: 20
        type: string
      name:
        maxLength: 128
        type: string
      priority:
        minimum: 0
        type: integer
      scope_code:
        maxLength: 128
        type: string
      scope_type:
        maxLength: 32
        type: string
    required:
    - model_id
    type: object
  schema.PolicyInvocation:
    properties:
      created_at:
        type: string
      creator:
        type: string
      description:
        type: string
      enabled:
        type: integer
      fallback_policy:
        type: string
      id:
        type: string
      model_id:
        type: string
      modifier:
        type: string
      name:
        type: string
      priority:
        type: integer
      retry_policy:
        type: string
      scope_code:
        type: string
      scope_type:
        type: string
      type:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
  schema.PolicyInvocationForm:
    properties:
      created_at:
        description: Create timestamp
        type: string
      creator:
        description: Creator
        type: string
      description:
        description: Details
        type: string
      enabled:
        description: Enabled
        type: integer
      fallback_policy:
        allOf:
        - $ref: '#/definitions/schema.FallbackPolicy'
        description: Fallback policy
      id:
        type: string
      model_id:
        description: Owner model ID; empty means template
        maxLength: 20
        type: string
      modifier:
        description: Modifier
        type: string
      name:
        description: Policy name
        maxLength: 128
        type: string
      priority:
        description: Priority
        minimum: 0
        type: integer
      retry_policy:
        allOf:
        - $ref: '#/definitions/schema.RetryPolicy'
        description: Retry policy
      scope_code:
        description: Scope code
        maxLength: 128
        type: string
      scope_type:
        description: Scope type
        maxLength: 32
        type: string
      type:
        description: Invocation type (failfast | failover)
        maxLength: 64
        type: string
      updated_at:
        description: Update timestamp
        type: string
      version:
        description: Version
        type: integer
    required:
    - name
    - type
    type: object
...
  schema.RetryPolicy:
    properties:
      backoff_type:
        description: 退避类型 (e.g. "fixed", "exponential")
        type: string
      base_ms:
        description: 退避间隔 (毫秒)
        type: integer
      code_policy:
        allOf:
        - $ref: '#/definitions/schema.ErrorParserPolicy'
        description: 错误码解析策略
      connect_timeout:
        description: 建立连接超时 (毫秒)
        type: integer
      error_codes:
        description: 需要重试的错误码/状态码列表
        items:
          type: string
        type: array
      error_messages:
        description: 需要重试的错误消息列表
        items:
          type: string
        type: array
      idle_timeout:
        description: 读空闲超时 (毫秒)
        type: integer
      message_policy:
        allOf:
        - $ref: '#/definitions/schema.ErrorParserPolicy'
        description: 错误消息解析策略
      retry:
        description: 重试次数
        type: integer
      total_timeout:
        description: 请求总超时 (毫秒)
        type: integer
      ttft_timeout:
        description: 首字超时 (毫秒)
        type: integer
      version:
        description: 版本标识 (保留)
        type: integer
    type: object
...
  util.ResponseResult:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_tokenlive_tokenlive-admin_pkg_errors.Error'
      success:
        type: boolean
      total:
        type: integer
    type: object
info:
  contact: {}
  description: An admin control center for tokelvie.
  title: tokenlive-admin
  version: v1.0.0
paths:
...
  /api/v1/policy/policy-circuit-breaks:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.PolicyCircuitBreak'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Query policy circuit break list
      tags:
      - PolicyCircuitBreakAPI
    post:
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.PolicyCircuitBreakForm'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.PolicyCircuitBreak'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Create policy circuit break record
      tags:
      - PolicyCircuitBreakAPI
  /api/v1/policy/policy-circuit-breaks/{id}:
    delete:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Delete policy circuit break record by ID
      tags:
      - PolicyCircuitBreakAPI
    get:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.PolicyCircuitBreakForm'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Get policy circuit break record by ID
      tags:
      - PolicyCircuitBreakAPI
    put:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.PolicyCircuitBreakForm'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Update policy circuit break record by ID
      tags:
      - PolicyCircuitBreakAPI
  /api/v1/policy/policy-circuit-breaks/{id}/copy-to-model:
    post:
      parameters:
      - description: template id
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.PolicyCopyToModelForm'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.PolicyCircuitBreak'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Copy policy circuit break template to model
      tags:
      - PolicyCircuitBreakAPI
  /api/v1/policy/policy-invocations:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.PolicyInvocation'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Query policy invocation list
      tags:
      - PolicyInvocationAPI
    post:
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.PolicyInvocationForm'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.PolicyInvocation'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Create policy invocation record
      tags:
      - PolicyInvocationAPI
  /api/v1/policy/policy-invocations/{id}:
    delete:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Delete policy invocation record by ID
      tags:
      - PolicyInvocationAPI
    get:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.PolicyInvocationForm'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Get policy invocation record by ID
      tags:
      - PolicyInvocationAPI
    put:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.PolicyInvocationForm'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Update policy invocation record by ID
      tags:
      - PolicyInvocationAPI
  /api/v1/policy/policy-invocations/{id}/copy-to-model:
    post:
      parameters:
      - description: template id
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.PolicyCopyToModelForm'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.PolicyInvocation'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Copy policy invocation template to model
      tags:
      - PolicyInvocationAPI
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"